set routing on fetchModulesAndSetRouting func
