+++ frontend/src/app/metadataModule/datasets/dataset-form.component.ts
@@ -50,10 +50,6 @@ export class DatasetFormComponent implements OnInit {
     this.form = this.datasetFormS.form;
 
     this._dfs.getTaxaBibList().subscribe(d => this.taxaBibList = d);
-    // get Modules
-    if (!this.moduleService.getModules()) {
-      this.moduleService.fetchModules();
-    }
 
     this.acquisitionFrameworks = this._dfs.getAcquisitionFrameworksList();
   }
